Update workaround
The workaround is incomplete with the private variables extracted to a separate file.

The underscore in the init file name causes it to load before OMF, which causes the new lines to not print. Naming it `pure_init.fish` loads it after `pure.fish`, which mimics the case when the variables weren’t separated.