Use `git name-rev` as a fallback for branch name
This would replace the current usage of `git show-ref`, which displays
commit hash when in detached HEAD, and bring it in line with how the Zsh
version displays branches.

When .git/HEAD is a hash and not a ref (e.g. when you `git checkout
HEAD~#` where # is some number of commits behind HEAD), it shows only
the hash.

One oddity to this approach is, if you have many branches or refs
containing common commits, the branch can show as `stash~2`,
`pullreqs/3071~1`, or `fix-colored-yes-perf~1`, for example. Because
this method is more descriptive, I consider this acceptable and better
than the current behavior of only showing the commit hash when in
detached HEAD.